fix: clear clickedNodeId on workflow select; reframe test as store unit
Addresses CodeRabbit review on PR #2799:
- setWorkflowSelected(true) now also clears clickedNodeId, preventing a
  stale value from re-opening the drawer on the next mousedown.
- Renames StudioNodeDrawer.integration.test.ts to a store unit test
  (workflowStoreCore.clickedNodeId.unit.test.ts) and removes the
  duplicated drawer predicate. The file now pins the store contract
  only; the drawer's render-time gating is covered separately.
